backend/requirements.txt
crewai>=1.11.0
crewai-tools>=0.41.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=13.0
